Precise integer type in depth function

+++ source/blender/blenkernel/BKE_grease_pencil.hh
@@ -229,7 +229,7 @@ class TreeNode : public ::GreasePencilLayerTreeNode {
   /**
    * \returns the number of non-null parents of the node.
    */
-  int depth() const;
+  int64_t depth() const;
 };
 
 /**
